ajout barre navig, I18n, modfi main.index

+++ webapp/main/classes/compte.py
@@ -8,7 +8,7 @@ class Compte(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     rib = db.Column(db.String(20), index=True, unique=True)
     created_at = db.Column(db.DateTime, index=True, default=datetime.utcnow)
-    solde = db.Column(db.Integer)
+    solde = db.Column(db.Float)
     titulaire_id = db.Column(db.Integer, db.ForeignKey('client.id'))
     discriminator = db.Column('type', db.String(50))
     __mapper_args__ = {'polymorphic_on': discriminator}
